catalog/requirements.txt
aniso8601==9.0.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
flake8==3.9.2
Flask==2.0.1
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.26.0
mccabe==0.6.1
mysql-connector-python==8.0.25
pymysql
redis>=4.0.0
protobuf==3.17.2
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.17
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
zipp==3.4.1
